requirements.txt
-e ../bigbee-remote_models
pytest
alembic
psycopg2-binary
python-dotenv
pyyaml